Fix normals in Blender exporter and viewer app
The correct vertices and indices are now written out by the Blender
exporter, depending on whether the triangle is flat or smooth
shaded. Flat shading means that vertices can not be re-used, as
their normals will be different. Smooth shading does allow vertex
re-usage via the index buffer if the normals for all vertices are
identical.

The viewer application loads and displays the normals correctly.
This also required the addition of a depth buffer to render the
geometry properly.