Use single timeline instead of multiple animations
Remove support for multiple animations and use a single timeline
with keyframes instead, which is more common and slightly
simplifies the file format.

Also clean up the transformations used for bone data in the Blender
exporter to not rely on a specific skinned mesh transform. It now
uses the transform of the actual armature object instead to
calculate the inverse bind pose matrix of a bone, and calculates
the transform between a bone in bind and animated pose directly in
armature space.