Add material module to converter app
This module correctly converts and channel-packs
GLTF/GLB textures to the AEM format and is
responsible for writing out the image buffer,
textures and materials.

Debug and customization options where added to the
viewer app: It is now possible to view individual
contributions to the final frame (base color,
vertex normals, normal maps, roughness and
metalness values and much more) and to customize
both lighting and tonemapping.