Improve movement code on ramps in showcase appasda
It is now possible to properly go up and down
ramps without slowing or gliding back down. This
was implemented with a two-phased approach where
horizontal and vertical movement is handled
separately.

Additionally, a second directional light was
added to slightly improve the lighting in shaded
areas. No clip mode was added to quickly move the
player around in levels. And a small glitch got
fixed where the player's angle would snap after
closing the debug window.