Add more hitboxes to enemy in showcase app asdasdasdasdas
Added upper and lower torso to the head hitbox for
the enemy and added a health system where a single
headshot is lethal but it takes several hits on
the other areas to kill the enemy.

Also added hurt sounds for the enemy, added sparks
to the headshot particle effects and cleaned up
some calls to a camera function.